Restructure giveaway components
