+++ packages/react-app/src/components/giveaways/details/PDFDocument.tsx
@@ -11,10 +11,10 @@ import {
   View,
 } from '@react-pdf/renderer';
 
-import CalendarIcon from '../../assets/CalendarIcon.png';
-import DownloadIcon from '../../assets/Download.png';
-import TrophyIcon from '../../assets/TrophyIcon.png';
-import { Giveaway } from '../../lib/types';
+import CalendarIcon from '../../../assets/CalendarIcon.png';
+import DownloadIcon from '../../../assets/Download.png';
+import TrophyIcon from '../../../assets/TrophyIcon.png';
+import { Giveaway } from '../../../lib/types';
 
 const styles = StyleSheet.create({
   section: {
