+++ packages/react-app/src/pages/details.tsx
@@ -3,8 +3,8 @@ import { useLoaderData, useRevalidator } from 'react-router-dom';
 
 import { Stack } from '@mui/material';
 
-import GiveawayAsideContent from '../components/giveaways/AsideContent';
-import GiveawayMainContent from '../components/giveaways/MainContent';
+import GiveawayAsideContent from '../components/giveaways/details/AsideContent';
+import GiveawayMainContent from '../components/giveaways/details/MainContent';
 import SubHeader from '../components/SubHeader';
 import { GetParticipants } from '../lib/queryClient';
 import { Giveaway, ParticipationState } from '../lib/types';
